requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
distro==1.9.0
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
langdetect==1.0.9
multidict==6.1.0
numpy==2.1.2
openai==1.52.0
playsound==1.2.2
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
tiktoken==0.8.0
tk==0.1.0
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
wavio==0.0.9
yarl==1.17.0
